# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SocialApp — unified social network + messaging platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      tags: [Posts]
      operationId: getFeed
      summary: Get social feed (following + own posts)
      parameters:
        - name: cursor
          in: query
          required: false
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paginated feed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListPage'
    post:
      tags: [Posts]
      operationId: createPost
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /posts/explore:
    get:
      tags: [Posts]
      operationId: getExplorePosts
      summary: Explore trending/recent public posts
      parameters:
        - name: cursor
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Paginated explore posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListPage'
  /posts/{postId}:
    get:
      tags: [Posts]
      operationId: getPost
      summary: Get a single post
      parameters:
        - name: postId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [Posts]
      operationId: updatePost
      summary: Update a post
      parameters:
        - name: postId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUpdate'
      responses:
        '200':
          description: Updated post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
    delete:
      tags: [Posts]
      operationId: deletePost
      summary: Delete a post
      parameters:
        - name: postId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Deleted
  /posts/{postId}/like:
    post:
      tags: [Posts]
      operationId: likePost
      summary: Like a post
      parameters:
        - name: postId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Liked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LikeResult'
    delete:
      tags: [Posts]
      operationId: unlikePost
      summary: Unlike a post
      parameters:
        - name: postId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Unliked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LikeResult'
  /posts/{postId}/save:
    post:
      tags: [Posts]
      operationId: savePost
      summary: Save a post
      parameters:
        - name: postId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveResult'
    delete:
      tags: [Posts]
      operationId: unsavePost
      summary: Unsave a post
      parameters:
        - name: postId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Unsaved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveResult'
  /posts/saved:
    get:
      tags: [Posts]
      operationId: getSavedPosts
      summary: Get my saved posts
      parameters:
        - name: cursor
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Saved posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListPage'

  # ── Comments ──────────────────────────────────────────────────────────────
...
  /stories:
    get:
      tags: [Stories]
      operationId: getStories
      summary: Get active stories from following
      responses:
        '200':
          description: Story groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StoryGroup'
    post:
      tags: [Stories]
      operationId: createStory
      summary: Create a story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoryInput'
      responses:
        '201':
          description: Story created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Story'
  /stories/{storyId}/view:
    post:
      tags: [Stories]
      operationId: viewStory
      summary: Mark a story as viewed
      parameters:
        - name: storyId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Viewed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
  /stories/{storyId}/viewers:
    get:
      tags: [Stories]
      operationId: getStoryViewers
      summary: Get viewers of a story
      parameters:
        - name: storyId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Viewers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListPage'

  # ── Chats ─────────────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true

    # ── Auth ────────────────────────────────────────────────────────────────
...
    UserSummary:
      type: object
      required: [id, username, displayName]
      properties:
        id: { type: string }
        username: { type: string }
        displayName: { type: string }
        avatarUrl: { type: ["string", "null"] }
        isFollowing: { type: boolean }
    UserListPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserSummary'
        nextCursor: { type: ["string", "null"] }
...
    Post:
      type: object
      required: [id, author, content, mediaUrls, likesCount, commentsCount, isLiked, isSaved, createdAt]
      properties:
        id: { type: string }
        author:
          $ref: '#/components/schemas/UserSummary'
        content: { type: string }
        mediaUrls:
          type: array
          items: { type: string }
        mediaType: { type: ["string", "null"], enum: [image, video, "null"] }
        likesCount: { type: integer }
        commentsCount: { type: integer }
        sharesCount: { type: integer }
        isLiked: { type: boolean }
        isSaved: { type: boolean }
        channelId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PostInput:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }
        mediaUrls:
          type: array
          items: { type: string }
        mediaType: { type: string, enum: [image, video] }
    PostUpdate:
      type: object
      properties:
        content: { type: string, minLength: 1 }
    PostListPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        nextCursor: { type: ["string", "null"] }
    LikeResult:
      type: object
      required: [liked, likesCount]
      properties:
        liked: { type: boolean }
        likesCount: { type: integer }
    SaveResult:
      type: object
      required: [saved]
      properties:
        saved: { type: boolean }

    # ── Comments ─────────────────────────────────────────────────────────────
...
    Story:
      type: object
      required: [id, author, mediaUrl, mediaType, viewsCount, expiresAt, createdAt]
      properties:
        id: { type: string }
        author:
          $ref: '#/components/schemas/UserSummary'
        mediaUrl: { type: string }
        mediaType: { type: string, enum: [image, video] }
        viewsCount: { type: integer }
        isViewed: { type: boolean }
        expiresAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    StoryGroup:
      type: object
      required: [user, stories, hasUnviewed]
      properties:
        user:
          $ref: '#/components/schemas/UserSummary'
        stories:
          type: array
          items:
            $ref: '#/components/schemas/Story'
        hasUnviewed: { type: boolean }
    StoryInput:
      type: object
      required: [mediaUrl, mediaType]
      properties:
        mediaUrl: { type: string, minLength: 1 }
        mediaType: { type: string, enum: [image, video] }

    # ── Chats ─────────────────────────────────────────────────────────────────
...